pyproject.toml
[tool.poetry]
name = "slapbot"
version = "0.2.0"
description = "Slap your friends in Telegram."
authors = ["Justin Arista Kasin <justin+github@aristakas.in>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^13.7"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
mypy = "^0.910"
pyre-extensions = "^0.0.22"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]